third_party/blink/web_tests/platform/generic/http/tests/devtools/sources/sources-panel-extension-names-expected.txt

The test verifies that extension names are resolved properly in navigator view.


Running: testAddExecutionContextBeforeFile
top
  FibExtension
    script.js
